Refactor build files
- move replays under src/3rdparty/replay
- remove system libuade support
- cleanup *FLAGS usage
- create library for each subproject to avoid compiling things many times
