+++ .gitignore
@@ -26,16 +26,17 @@ INSTALL
 *.la
 .deps/
 .libs/
+*.a
+*.dSYM
 stamp-h1
 config.h
 config.h.in
 config.h.in~
 .dirstamp
-*.dSYM
 m4
 bak
+audacious-uade-*
 *.tar.bz2
 debian/changelog
 build-deb
 build
-lib3rdparty.a
