+++ configure.ac
@@ -1,5 +1,5 @@
 AC_PREREQ(2.59)
-AC_INIT([audacious-uade], m4_normalize(m4_include([VERSION])))
+AC_INIT([audacious-uade], m4_normalize(m4_include([VERSION])), [], [], [https://github.com/mvtiaine/audacious-uade])
 AC_PROG_CXX
 AX_CXX_COMPILE_STDCXX_17
 
@@ -41,12 +41,6 @@ AM_CONDITIONAL([BUILD_WIN32], [test "$build_windows" = "yes"])
 AM_CONDITIONAL([BUILD_MACOS], [test "$build_macos" = "yes"])
 AM_CONDITIONAL([BUILD_BSD], [test "$build_bsd" = "yes"])
 
-AC_ARG_WITH([system-libuade],
-  AS_HELP_STRING([--with-system-libuade],[Use system libuade instead of bundled one]),
-  [system_libuade="yes"])
-AS_IF([test "$system_libuade" = "yes"], [AC_DEFINE([SYSTEM_LIBUADE], [1], [Use system libuade])])
-AM_CONDITIONAL([SYSTEM_LIBUADE], [test "$system_libuade" = "yes"])
-
 PKG_CHECK_MODULES([AUDACIOUS],[audacious >= 3.8],
   [ac_audacious_plugindir=`$PKG_CONFIG --variable=plugin_dir audacious`],
   [ac_audacious_plugindir=;])
@@ -63,52 +57,50 @@ AC_ARG_WITH([audacious-plugindir],
 AC_SUBST(AUDACIOUS_PLUGINDIR)
 AC_MSG_RESULT($AUDACIOUS_PLUGINDIR)
 
+if test "$build_linux" = "yes"
+then
+  PKG_CHECK_MODULES([LIBBSD],[libbsd],
+    [ac_libbsd_libs=`$PKG_CONFIG --libs libbsd`],
+    [ac_libbsd_libs=;])
+fi
+
+CFLAGS="-fPIC -fvisibility=hidden -fdata-sections -ffunction-sections ${CFLAGS}"
+CXXFLAGS="-fno-rtti -Wall ${CFLAGS}"
+if test "$build_macos" = "yes"
+then
+  LDFLAGS="-Wl,-dead_strip -Wl,-undefined,error ${LDFLAGS}"
+  SHREXT="-shrext .dylib"
+  AC_SUBST([SHREXT])
+else
+  LDFLAGS="-Wl,--gc-sections -Wl,--as-needed -Wl,--exclude-libs=ALL -Wl,--discard-all -Wl,--no-undefined -Wl,-O1 ${LDFLAGS}"
+fi
+
+uade_prefix=${prefix}
+if test "$uade_prefix" = "NONE"
+then
+  uade_prefix=/usr/local
+fi
+AX_SUBDIRS_CONFIGURE([uade], [
+  [--without-uade123], [--without-uadesimple], [--without-write-audio],[--without-uadefs],
+  [--prefix=\${uade_prefix}/share/audacious-uade],
+  [--bindir=/dev/null], [--mandir=/dev/null], [--sharedir=/dev/null],
+  [--bencode-tools-prefix=/dev/null], [--libzakalwe-prefix=/dev/null],
+  [CFLAGS=${CFLAGS} -DDISABLE_BENCODETOOLS -DDISABLE_ZAKALWE -Wno-format -Wno-format-security],
+],[])
+UADE_CFLAGS="-I\${top_srcdir}/uade/src/frontends/include -I\${top_builddir}/uade/src/frontends/include"
+AC_SUBST([UADE_CFLAGS])
+# TODO shutup libtool warning:
+# *** Warning: Linking the shared library uade.la against the
+# *** static library uade/src/frontends/libuade/libuade.a is not portable!
+UADE_LIBS="\${top_builddir}/uade/src/frontends/libuade/libuade.a"
+AC_SUBST([UADE_LIBS])
+
 AC_CONFIG_HEADERS([src/config.h])
 AC_CONFIG_FILES(Makefile
                 src/Makefile
+                src/3rdparty/Makefile
                 src/converter/Makefile
+                src/player/Makefile
                 src/songend/Makefile)
-
-AC_MSG_CHECKING([whether to use system libuade])
-if test "$system_libuade" = "yes"
-then
-  AC_MSG_RESULT(yes)
-  PKG_CHECK_MODULES([UADE],[libuade >= 3.0],
-    [ac_uade_datadir=`$PKG_CONFIG --variable=data_directory libuade`],
-    [ac_uade_datadir=;])
-  AC_SUBST(UADE_CFLAGS)
-  AC_SUBST(UADE_LIBS)
-  AC_SUBST(UADE_SYSTEM_DATADIR, $ac_uade_datadir)
-else
-  AC_MSG_RESULT(no)
-  if ! test -d "$srcdir/uade"; then
-    AC_MSG_ERROR([Could not find uade submodule. Please retry after running: git submodule update --init --recursive])
-  fi
-  if test "$build_linux" = "yes"
-  then
-    PKG_CHECK_MODULES([LIBBSD],[libbsd],
-      [ac_libbsd_libs=`$PKG_CONFIG --libs libbsd`],
-      [ac_libbsd_libs=;])
-    export LDFLAGS="$ac_libbsd_libs ${LDFLAGS}"
-  fi
-
-  export CFLAGS="-DDISABLE_BENCODETOOLS -DDISABLE_ZAKALWE -fvisibility=hidden -fdata-sections -ffunction-sections ${CFLAGS} -Wno-format -Wno-format-security"
-  uade_prefix=${prefix}
-  if test "$uade_prefix" = "NONE"
-  then
-    uade_prefix=/usr/local
-  fi
-  AX_SUBDIRS_CONFIGURE([uade], [
-    [--without-uade123], [--without-uadesimple], [--without-write-audio],[--without-uadefs],
-    [--prefix=\${uade_prefix}/share/audacious-uade],
-    [--bindir=/dev/null], [--mandir=/dev/null], [--sharedir=/dev/null],
-    [--bencode-tools-prefix=/dev/null], [--libzakalwe-prefix=/dev/null],
-  ], [])
-  UADE_CFLAGS="-I\${top_srcdir}/uade/src/frontends/include -I\${top_builddir}/uade/src/frontends/include"
-  # TODO shutup libtool warning:
-  # *** Warning: Linking the shared library uade.la against the
-  # *** static library uade/src/frontends/libuade/libuade.a is not portable!
-  UADE_LIBS="\${top_builddir}/uade/src/frontends/libuade/libuade.a"
-fi
 
 AC_OUTPUT
