+++ src/converter/Makefile.am
@@ -1,8 +1,9 @@
+AUTOMAKE_OPTIONS = subdir-objects
+noinst_LIBRARIES = libconverter.a
 noinst_PROGRAMS = uade_convert
-uade_convertdir = $(prefix)/bin
-uade_convert_SOURCES = \
+
+libconverter_a_SOURCES = \
     big_endian.h \
-    converter_main.cc \
     converter.h \
     converter.cc \
     common_med.h \
@@ -11,3 +12,6 @@ uade_convert_SOURCES = \
     ../3rdparty/macos_endian.h \
     ../3rdparty/SimpleBinStream.h \
     ../3rdparty/proplayer.h
+
+uade_convert_LDADD = libconverter.a
+uade_convert_SOURCES = converter_main.cc
