+++ src/hacks.cc
@@ -15,13 +15,9 @@
 
 #include "config.h"
 extern "C" {
-#if SYSTEM_LIBUADE
-#include <uade/uade.h>
-#else
 #include "../uade/src/frontends/include/uade/options.h"
 #include "../uade/src/frontends/include/uade/uadeconfstructure.h"
 #include "../uade/src/frontends/include/uade/uade.h"
-#endif
 }
 #include "common.h"
 #include "extensions.h"
