+++ src/player/Makefile.am
@@ -1,0 +1,12 @@
+AUTOMAKE_OPTIONS = subdir-objects
+noinst_LIBRARIES = libplayer.a
+
+libplayer_a_SOURCES = \
+    foreach.h \
+    player.h \
+    player.cc \
+    player_dbm.cc \
+    player_hvl.cc
+
+libplayer_a_DEPENDENCIES = \
+    ../3rdparty/lib3rdparty.a
