+++ src/plugin.cc
@@ -15,13 +15,9 @@
 
 #include "config.h"
 extern "C" {
-#if SYSTEM_LIBUADE
-#include <uade/uade.h>
-#else
 #include "../uade/src/frontends/include/uade/options.h"
 #include "../uade/src/frontends/include/uade/uadeconfstructure.h"
 #include "../uade/src/frontends/include/uade/uade.h"
-#endif
 }
 #include "common.h"
 #include "extensions.h"
@@ -316,17 +312,13 @@ public:
     static constexpr PluginInfo info = {
         "UADE Plugin",
         "audacious-uade",
-        "Audacious UADE plugin " VERSION "\n"
+        "Audacious UADE plugin " PACKAGE_VERSION "\n"
         "Written by Matti Tiainen <mvtiaine@cc.hut.fi>\n"
-        "https://github.com/mvtiaine/audacious-uade\n"
+        PACKAGE_URL"\n"
         "\n"
         "UADE: https://zakalwe.fi/uade/\n"
         "\n"
-#if SYSTEM_LIBUADE
-        "Using system libuade " UADE_VERSION "\n"
-#else
         "Using bundled libuade " UADE_VERSION "\n"
-#endif
         "\n"
         "Simplistic Binary Streams 1.0.3\n"
         "Copyright (C) 2014-2019, Wong Shao Voon\n"
