+++ src/uade_common.cc
@@ -63,16 +63,9 @@ struct uade_state *create_uade_probe_state(int freq) {
 #if DEBUG_TRACE
     uade_config_set_option(uc, UC_VERBOSE, nullptr);
 #endif
-    // use our uade.conf, song.conf and contentdb even with system libuade
+    // use our uade.conf, song.conf and contentdb
     uade_config_set_option(uc, UC_BASE_DIR, UADEDATADIR);
-#if SYSTEM_LIBUADE
-    // make sure to use system version for these
-    uade_config_set_option(uc, UC_UAE_CONFIG_FILE, UADE_CONFIG_BASE_DIR "/uaerc");
-    uade_config_set_option(uc, UC_SCORE_FILE, UADE_CONFIG_BASE_DIR "/score");
-    uade_config_set_option(uc, UC_UADECORE_FILE, UADE_CONFIG_UADE_CORE);
-#else
     uade_config_set_option(uc, UC_NO_CONTENT_DB, nullptr);
-#endif
     uade_config_set_option(uc, UC_ONE_SUBSONG, nullptr);
     uade_config_set_option(uc, UC_TIMEOUT_VALUE, "-1");
     uade_config_set_option(uc, UC_DISABLE_TIMEOUTS, nullptr);
