+++ src/uade_common.h
@@ -6,13 +6,9 @@
 
 #include "config.h"
 extern "C" {
-#if SYSTEM_LIBUADE
-#include <uade/uade.h>
-#else
 #include "../uade/src/frontends/include/uade/options.h"
 #include "../uade/src/frontends/include/uade/uadeconfstructure.h"
 #include "../uade/src/frontends/include/uade/uade.h"
-#endif
 }
 #include <cassert>
 #include <pthread.h>
